+++ internal/message/models/models.go
@@ -12,18 +12,18 @@ import (
 
 // Message represents a message in the database.
 type Message struct {
-	ID             int64              `db:"id" json:"id"`
+	ID             int                `db:"id" json:"id"`
 	CreatedAt      time.Time          `db:"created_at" json:"created_at"`
 	UpdatedAt      time.Time          `db:"updated_at" json:"updated_at"`
 	UUID           string             `db:"uuid" json:"uuid"`
 	Type           string             `db:"type" json:"type"`
 	Status         string             `db:"status" json:"status"`
-	ConversationID int64              `db:"conversation_id" json:"conversation_id"`
+	ConversationID int                `db:"conversation_id" json:"conversation_id"`
 	Content        string             `db:"content" json:"content"`
 	ContentType    string             `db:"content_type" json:"content_type"`
 	Private        bool               `db:"private" json:"private"`
 	SourceID       null.String        `db:"source_id" json:"-"`
-	SenderID       int64              `db:"sender_id" json:"sender_id"`
+	SenderID       int                `db:"sender_id" json:"sender_id"`
 	SenderType     string             `db:"sender_type" json:"sender_type"`
 	InboxID        int                `db:"inbox_id" json:"-"`
 	Meta           string             `db:"meta" json:"meta"`
